resume_optimizer/requirements.txt
streamlit==1.36.0
PyPDF2>=3.0.0
python-docx>=0.8.11
fpdf2>=2.7.9
scikit-learn>=1.3.0
nltk>=3.8.1
groq>=0.16.0
